+++ jobbergate-api/dev_tools/entrypoint.sh
@@ -1,6 +1,6 @@
 #!/bin/bash
 
+set -e
+cd /app
 poetry run dev-tools db upgrade
-poetry run dev-tools dev-server --port=8000
-
-tail -f /dev/null
+poetry run dev-tools dev-server --port=80