+++ jobbergate-cli/CHANGELOG.rst
@@ -7,6 +7,8 @@ This file keeps track of all notable changes to jobbergate-cli
 Unreleased
 ----------
 - Added error details to configuration error report
+- Added ``OIDC_USE_HTTPS`` setting to allow non-https OIDC hosts
+- Removed cluster validation from job-submission due to reliance on external cluster registry
 
 3.2.5-alpha.0 -- 2022-09-15
 ---------------------------